Simplify the basic example and make a timing example
